requirements.txt
# Core Dependencies
Flask==3.1.0
flask-cors==5.0.1
numpy==1.26.0
pydantic==2.10.6
sentencepiece==0.2.0
openai==1.66.3
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
accelerate==1.5.2
diffusers==0.32.2
SwissArmyTransformer==0.4.12

# Media Processing
opencv-python==4.11.0.86
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
scikit-video==1.1.11
basicsr==1.4.2
realesrgan==0.3.0

# Web & UI
gradio==5.21.0

# Additional useful tools (from working container)
protobuf==6.30.1
pillow==10.4.0
tqdm==4.67.1
scikit-image==0.25.2
matplotlib==3.10.1
requests==2.32.3

# Environment & Utility
python-dotenv==1.0.1
pymongo==4.7.2  # Ensure latest compatible with Python version
pytz==2024.1
